refactor: rename create to details
