+++ src/app/proposals/new/details/activation/page.tsx
