+++ src/app/proposals/new/details/components/BaseProposalFields.tsx
