+++ src/app/proposals/new/details/components/index.ts
