+++ src/app/proposals/new/details/deactivation/page.tsx
