+++ src/app/proposals/new/details/layout.tsx
