+++ src/app/proposals/new/details/schemas/ActivationProposalSchema.ts
