+++ src/app/proposals/new/details/schemas/BaseProposalSchema.ts
