+++ src/app/proposals/new/details/schemas/DeactivationProposalSchema.ts
