+++ src/app/proposals/shared/types.ts
@@ -1,9 +1,9 @@
 import Big from '@/lib/big'
 import { DecodedData } from '@/app/proposals/shared/utils'
 import { ProposalCategory, ProposalState } from '@/shared/types'
-import { type GrantProposal } from '../new/create/schemas/GrantProposalSchema'
-import { type ActivationProposal } from '../new/create/schemas/ActivationProposalSchema'
-import { type DeactivationProposal } from '../new/create/schemas/DeactivationProposalSchema'
+import { type GrantProposal } from '../new/details/schemas/GrantProposalSchema'
+import { type ActivationProposal } from '../new/details/schemas/ActivationProposalSchema'
+import { type DeactivationProposal } from '../new/details/schemas/DeactivationProposalSchema'
 export interface Proposal {
   votes: {
     againstVotes: Big
